Fix "wobbly" vertexes
When the player stands on a vertex with a change in height in Vanilla
DOOM, the vertex "wobbles". A fix was put in R_ScaleFromGlobalAngle() by
changing max from 64 * FRACUNIT to 256 * FRACUNIT. This improved the
problem, but not entirely. A value of 1024 * FRACUNIT is much better,
but previously would cause overflows in more complicated scenes. With
limits now removed, this value can be used.