Fix hitscan checks
From ZDoom: "...two bugs which makes hitscan attacks less likely to hit.
The first is that it is a monster's cross-section, rather than its
bounding box, that is used to check for impact; this makes attacks with
a limited range (especially melee attacks) unlikely to hit very wide
monsters. The second is the blockmap bug: if an actor crosses block
boundaries and its center is in a different block than the one in which
the impact happens, then there is no collision at all, letting attacks
pass through it harmlessly."