Save correct skill level
A long time ago I attempted to fix the "demon speed bug" (see
http://doomwiki.org/wiki/Demon_speed_bug) by simply saving d_skill
instead of gameskill to the savegame header. This causes the wrong skill
level to be set in savegames sometimes when -fast is used (and perhaps
in other instances). I'll need to look at other solutions to fix the
demon speed bug.