Clip pickups using real radius rather than 20
All pickups in Vanilla DOOM have a radius of 20 * FRACUNIT. This is fine
when checking when the player picks them up, but that same radius is
used when checking those pickups against lines, and can cause small
things (such as MT_CLIP) to hang midair. This fix uses 20 * FRACUNIT
when picking up, and a more accurate radius elsewhere.