Display TEKWALL1 texture the same as Vanilla DOOM
R_DrawColumnInCache() in r_data.c contains a fix that draws multipatch
textures correctly. This improves the look of STEP2 on the ceiling of
the room with the imps in MAP01 of DOOM II: Hell On Earth. But it also
changes the look of TEKWALL1. This commit makes TEKWALL1 a special
instance, and doesn't apply the fix to this texture. Example pics to
follow in comments...