Put savegames in individual folders with WAD name
Since the config folder is now the same folder as the executable, by
default, D_SetSaveGameDir() puts all savegames in one place. This commit
changes it so savegames are still put in individual folders based on the
name of the WAD used (the same as what happened when the config folder
was AppData\Local\DOOM RETRO).