Blood splats don't need thinkers
Blood splats are spawned where blood hits the ground, so they don't
move, and they start being removed when the "bloodsplats" limit is
reached, not when mobj->tics is reached. So there's up to 1024 thinkers
that aren't necessary. I'm hoping this provides a major performance
boost and I can increase that limit...