Implement unlimited blood splats by default
After some rigorous playtesting, it would seem that after the several
recent optimizations I have made towards blood splats and other mobj's
in general, there can now be a significantly large amount of blood
splats with no noticeable slowdown. Make this the default, but
"bloodsplats" can still be changed to a value 0 to 32768 in default.cfg.