Vibrate XBOX 360 controller when firing weapon
A bit more work to go, but now force feedback has been implemented for
XBOX 360 controllers when firing weapons. Still need to experiment to
get the right amount of vibration for each weapon, and also vibrate when
chainsaw is idle. Some optimizations have been made to the existing
vibration code as well.