Initial implementation of random pitch shifting
Randomly shifts the pitch of all sounds in game, as originally intended
in the DOOM source, and as now occurs in the official XBLA and BFG
Edition versions of DOOM. This is just the start. I have further ideas
regarding this feature. I'll at least make it an option. And off by
default probably. Taken from Strawberry DOOM.