Revamp berserk power-up palette effect
I've never been a fan of the berserk power-up palette effect. Firstly,
as it fades out, the jump between the shades of red is distracting.
Secondly, even today it confuses people who think the power-up only
lasts for as long as the screen is red, rather than to the end of the
level. This commit changes this so that the screen is a constant tint of
red while the player has the berserk power-up and their fist selected.