Remove hack that notifies statbar if automap on/off
From Eternity Engine, 09/29/04: "Finally got around to removing the
weird hack that Dave Taylor put into ST_Responder which allowed the
automap to notify the status bar when the game was changing into or out
of automap mode. The code to do that did NOT belong in ST_Responder, and
the notification messages were being sent in via hacked event_t
structures that contained values which were assumed to not alias any
valid input. This made an unnecessary AND-op and branch in ST_Responder
on every single input event, which is gross considering the nature of
the hack. Created a function called ST_AutomapEvent which the automap
code now calls instead of ST_Responder."