Fix check to see if player is on an object
The original code, taken from
https://github.com/chocolate-doom/chocolate-doom/blob/master/src/hexen/p_mobj.c#L1089,
incorrectly checks for MF2_ONMOBJ in flags when it seems it should be
flags2. (The same problem appears in the original Hexen source as well.)
Changing it to flags2 is a bit redundant since you're clearing the flag
the very next line, so I figure the best solution is to remove the test
entirely. We'll see how this goes...