Simplify check for unsticking stuck monsters
Simplify code from commit 9224eb305638f5682edc33f13c4c5c1699464037. As
brought to my attention by Jeff Doggett, tmz is only used when
telestomping. It isn't necessary to check it. Tested this using MAP24 in
NOVA.WAD. Cacodemons tended to get stuck in that map, as there are quite
a lot of them, but not anymore!