Fix use action in finale carrying over to next map
As an example, if you pressed the space bar during the finale text at
the end of MAP06 in DOOM II, the text would skip, but the switch in
front of the player at the start of MAP07 would be triggered by that
press. This is fixed. Also, pressing a key will now display the finale
text in full, and another press will then advance. Thanks to Jon Krazov
for this one.