Restore monster targets when loading game
In Vanilla DOOM, monster targets are nullified when loading a game
because they no longer point to the targets, and instead point to the
wrong parts of memory. What this commit does is if the monster had a
target when the game was saved, their target is set to the player when
loading the game. This will break any monster infighting, but I feel its
preferable to monsters acting all surprised again when they see you.