Switch to chainsaw when starting new level
If the player has a chainsaw and a berserk power-up, and ends a level
with their fists selected, they start the next level with their fists.
This shouldn't happen, because the berserk power-up has been removed and
they have the chainsaw. This commit corrects this. Thanks to @plumsinus
for finding this!