Fix crash due to overflow
The change previously made to R_ScaleFromGlobalAngle() to "fix wobbly
vertexes", although improves rendering of edges of spans, can cause
overflows. Previously I thought this only caused occasional visual
anomalies, but in MAP03 of SCYTHE2.WAD, it causes a crash when you fall
down to the indoor water area and walk over the "cracks" in the floor
(linedefs 2400, 2403-2406 and 2408).