Improve wobbly line problem
Changing the max value in R_ScaleFromGlobalAngle() from 64 to 256
improves the wobbly line problem, but doesn't crash as described in
commit 9b7a767f6f239358d5377154c9f04db34a9220c8. A further idea may be
to choose a value (64, 128, 256, 512 or 1024) on a per-level basis,
dependent on the greatest height difference between 2 sectors in the
level. There's an old but interesting discussion at
http://www.doomworld.com/vb/source-ports/38425-solution-to-wobbly-line-problem-that-doesnt-involve-build-code/.
This will do for now.