Implement autorun feature
Implemented an "autorun" setting in doomretro.cfg, which is "false" by
default. If caps lock is turned on during the game, "autorun" is set to
"true" and caps lock is automatically turned on the next time the game
is started. Also, an "AUTORUN ON/OFF" message is now displayed when the
caps lock key is pressed.