Fix fire/use actions carrying over into game
When pressing fire or use actions to respawn after death when a game has
been quicksaved, and when using same actions to load a save game from
the menu, those actions no longer carry over through to the game. For
instance, activating the switch directly in front of the player in MAP07
of DOOM2.WAD. Thanks to Jon Krazov.