Don't draw teleport lines until triggered
Draw teleport lines as normal 2s lines in the automap until they are
walked on by the player. Then draw that line and all connected teleport
lines to that sector in dark red. Using the IDDT cheat will still show
all teleport lines in dark red regardless. This change was inspired by
the fact that all teleport lines in MAP30 of TNT.WAD were shown, which
gave away the "puzzle" towards the start of that map.