Correctly validate key settings (take two)
I was going about it all wrong in the previous commit. SDL_GetKeyName()
wasn't working like I wanted anyway. This time, I've noticed that the
key is already checked in LoadDefaultCollection(): if its < 0 or > 127,
its set to 0. Now, I set a key to -1 in this instance, and then check
again in M_CheckDefaults() to set the key to its default if necessary.