Implement drop shadows under monsters
Not exactly id Tech 5 or anything, but now monsters cast translucent
shadows. Looks quite effective, and unique. The process of getting this
working involves creating a drop shadow for each frame. This is a bit
laborious, so only the Zombieman is done in this commit, with the other
monsters to follow in the coming days.