Fix widescreen mode aspect ratio on 16:9 displays
Widescreen mode is ideal on a 16:10 display (the player's view without
the status bar is 16:10 and widescreen mode effectively zooms in on that
view). 16:9 displays are becoming more popular, so rather than
stretching the view horizontally to fill that extra width, the view's
width is calculated correctly to retain the 16:10 aspect ratio. This
will result in pillarboxes though (for example, pillarboxes of 85px for
a resolution of 1600×900).