Fix firing animation of Heavy Weapon Dude
This is a known discrepancy between the offsets of the two firing frames
of the Heavy Weapon Dude when facing to the right (CPOSE7 and CPOSE8).
This was meant to be fixed by the custom offsets in info.c as far back
as v1.0 of DOOM RETRO, but because of a stupid typo, it turns out it
hasn't been fixed all this time.