Remove corpses from map in Chex
The corpses of monsters in Chex are invisible, which means in DOOM RETRO
there's also an invisible shadow object left behind, and a thing
triangle is displayed in the automap. This commit is a small hack that
sets the tics of the last frame of each monster's death sequence to 0 so
both the corpse and shadow objects are removed.