Allow display of asterisk in save description
Previously a degree symbol replaced the asterisk in the default
character set for the benefit of being displayed in the message for the
IDMYPOS cheat. The degrees symbol is now a separate character. Minor
changes have also been made: '\n' will be replaced with nothing rather
than a space, and the degrees symbol will be replaced with nothing
rather than a space if a custom character set is used.