Allow widescreen mode when in a window
Setting the screen size (through the options menu or with the +/- keys)
to the two highest settings will now enable widescreen mode when in a
window. Previously, widescreen mode was only available when fullscreen
on widescreen monitors. In windowed mode, the window actually increases
in width. A bit unconventional, but the alternatives are to either just
crop the status bar, or have pillarboxes always present when not in
widescreen mode.
There's still more work to be done. Need to update the window resize
code, handle ALT+ENTER when in widescreen mode, and center the window
correctly.