Fix slowdown when many mobjs on moving sector
Thanks to Platinum Shell on the Doomworld forums for bringing this
problem to my attention: When there's a large amount of monsters (and
especially blood splats) on a moving sector, there's some seriously
nasty slowdown. An example of this is MAP30 in Doom II: put god mode on,
leave a few monsters to spawn and fight amongst themselves for a few
minutes, and then lower one of the floors and you'll see what I mean.
The changes I've made in this commit greatly speed up things, with no
adverse side effects as of yet.