Correct amount in message on ammo pickup
When player picks up ammo, the messages displayed are hardcoded to
indicate "one cell", "four shells", etc., but when the lowest and
highest skill levels are chosen, the player is actually given double the
amount. (Also, there's a difference between clips dropped by Zombiemen,
and clips that aren't.) This commit implements 4 new dehackable messages
to display the correct amount.