Allow change to gamepad left/right dead zones
As discussed in issue #58, have now exposed the gamepad's left and right
thumbstick's dead zones to be changed. The new settings are
gamepad_leftdeadzone and gamepad_rightdeadzone, with defaults of 7849
and 8689, respectively. Lowering these values will make the thumbsticks
more responsive, but depending on the gamepad (and each gamepad is
different), too low a value may cause one or both of the thumbsticks to
"travel" at rest.