Fix some shootable switches only working once
Since the first version of DOOM RETRO, a small hack has been in place
that fixes the switch sound from playing when shooting through the
doorway at the start of MAP18 in DOOM2.WAD. Unfortunately, this caused
shootable switches to only work once. In MAP03 of Memento Mori 2, a
switch needs to be shot on two occasions to progress through the map. So
now this hack is changed to only work in MAP18 of DOOM2.WAD. Thanks to
Platinum Shell. http://www.doomworld.com/vb/post/1328915.