Spawn corpse blood in circle rather than square
When the corpses setting in doomretro.cfg has the moreblood flag set,
blood splats are spawned around any decorative corpses when a map is
loaded. Now the blood splats are spawned in a circle about the corpses,
rather than a square which looks more natural. Also, the circle of blood
is randomly shifted slightly.