Default to DOOM2.WAD if no map data in PWAD
Previously, if no map data was present in the PWAD(s) selected in the
WAD selector, and no IWAD was selected with them, DOOM RETRO wouldn't
know what IWAD to load and give up. Now, DOOM RETRO will try to load
DOOM2.WAD as default in this instance. This commit exists to make it
easier to load Linguica's InstaDoom. :grin: :camera: