Bring back SDL1.2 code
Although the SDL2 code is coming along nicely, and will be used in the
next release of DOOM RETRO, SDL2 doesn't play nice with Raspberry Pi, as
brought to my attention in issue #94 by @sneakernets. So here it is back
again, #defined out as necessary, but this time, there are no separate
solution configurations for MSVC. As going through this, I also replaced
all instances of "#if[n]def" with "#if [!]defined(" since according to
https://msdn.microsoft.com/en-us/library/2a1b21sf.aspx they "are
provided only for compatibility with previous versions of the language.
The defined( identifier ) constant expression used with the #if
directive is preferred." A few other minor tweaks crept in as well.