Fix display of automap title if loaded using ZDL
For whatever reason, ZDL uses the wrong directory separator for Windows.
When constructing the automap title for PWADs without a map name in a
DEHACKED lump, DOOM RETRO uses the name of the PWAD taken from its path
using ExtractFileBase() in w_wad.c. Now it checks for '\' and '/'
separators indiscriminately.