Hack to fix sound volume if SDL crashed
On occasion SDL may crash, and when starting up DOOM RETRO again, the
sfx and music volumes are muted. Going into the sound volume menu and
adjusting the music volume brings the sound back, so this commit sets
the music volume twice at startup. Hopefully this will fix this issue.