Don't add thinkers for bloodsplats or shadows
Every bloodsplat and shadow has a thinker function. The function itself
is empty, and the position of the objects are handled only when
necessary elsewhere, but that's still a lot of thinkers. This means that
whenever the list of thinkers is checked, a majority of them have to be
skipped over to find a monster, etc. This commit removes the creation of
thinkers for bloodsplats and shadows completely, greatly increasing the
efficiency of functions that iterate through the list of thinkers every
tic (or whenever), such as P_RunThinkers() and P_LookForPlayers().
Saving/loading of mobjs required a rewrite as well.