Implement dpi-awareness
The game already runs fine on high-dpi display, thanks to SDL2, but the
screen resolution was displayed incorrectly in the console at startup.
Probably a better way of implementing this is to make the app dpi-aware
through a manifest, but it would seem this isn't supported when using
the v90 platform toolset in MSVC (and I want to keep that as it is so
the app continues to work in Linux under Wine).