Fix window not resizable after being fullscreen
If game starts in fullscreen mode, when switching to a window, the
window isn't resizable because the SDL_WINDOW_RESIZABLE flag hasn't been
applied at any point. So apply the flag at startup, whether the game is
in fullscreen or not. Using the flag in the call to
SDL_SetWindowFullscreen() wasn't the correct way.