Revamp console background and color scheme
Instead of using a background pic, blur and darken, consistent with the
menu background. Have different colors for BTSX's player messages still,
but the other colors are now the same regardless of IWAD. Mostly light
gray, with white for the input. The idea is to keep things generic and
clean. It also resembles Doom 3's console somewhat.