Set default zoom of every map the same in automap
The larger the map, the further zoomed out the map appears when entering
the automap the first time. For some very large maps, it makes the
automap useless without zooming in first, and "gives away" that the map
is huge, (which in my opinion, is something you should discover by
actually playing the map). What this commit does is initially set every
map the same, so it appears at the same zoom level as E1M1: Hangar does.
I did have this in previous versions of DOOM RETRO, but it was recently
removed in commit 3f84be38aa49fb10be7336b81c07643159406da9.