Various improvements to kill cmd
* Don't allow kill with no parms while player is already dead.
* When killing player, kill them for the amount of health they have
left. This means that the screen will flash red, and the health, etc. in
the HUD/status bar will be updated correctly. (Thanks to @JonKrazov for
raising this in issue #127).
* If the player is killed (or one or more monsters are) automatically
close the console, to see the immediate effects of your
suicide/genocide. (I'll work on automatically closing the console for
other commands as well...)