Still autoload previous savegame on player death even if level has changed
Previously, if the player saved their game, exited a level, and then
died on the following level, rather than the savegame from the previous
level autoloading, the player would pistol start on the new level. This
commit fixes this. Another idea I'm considering is if a player
progresses to the next level without cheating, the game is autosaved at
the start of the level...