Complete support for translucent midtextures
The problem in getting this working, as described in my lonely monologue
in commit 102d5c5e92072095cd2749ff9457f1032d481aa9, was because the
interpolation of sector heights was interfering with Killough's deep
water/fake ceiling effect. Also in this commit reordered the setup of
each level a bit.