Look for another line to use if 1st one is wrong side
Normally with a use action, the closest usable line within range in
front of the player is found. If the player is on the wrong side of that
line, now PTR_UseTraverse() will keep looking for the next usable line
that the player is on the right side of. (This fixes issues such as in
C1M5 of Freedoom1.wad: there's a lift very close to 2 doors. You either
can't open the door once the lift is lowered, or vice versa from the
bottom. Pic to follow in comments.)
Interestingly, this isn't addressed in PrBoom+. Works okay in ZDoom, but
it does things differently.