Highlight picked up health/ammo/armor in HUD
When health, ammo or armor is picked up by the player, the corresponding
element in the HUD will be highlighted briefly. The translucency of the
HUD has been reduced slightly to give better distinction between states,
but I haven't settled on the final highlight effect yet, so this may
change.