Don't always apply the long wall error fix
@linguica's otherwise excellent long wall error fix can cause graphic
anomalies to appear in some instances. Examples include the 3D bridge in
MAP01 of SCYTHE2.WAD, and under the vines at the player start in MAP01
of VALIANT.WAD. So, if the line is small enough, or a masked texture is
used, the old method will now be used instead.