Seed random numbers twice for each map
First, seed them before things are spawned onto the map using the number
of things in the map. This is so blood splats spawned around corpses are
the same each time for that map, but different between maps. Then once
all things are spawned, seed random numbers again using the time for
better randomness.