Randomize pitch per monster instead of per sound
When the r_randompitch cvar is on, instead of randomizing each sound
regardless of where it came from (which is what happens with the feature
in Doom 1.2, Heretic, Hexen, and afaik any source port that supports
this feature), randomize "per monster" instead.
This means that when a monster is spawned, all the sounds it makes are
randomized to the same pitch, rather than each sound it makes being
randomized to different pitches. I feel that this sounds more "natural",
although this feature will still be off by default I think.