Make Direct3D the default renderer
There appears to be a problem with tearing using the OpenGL renderer in
SDL2 when fullscreen. I'm guessing it has something to do with how it's
implemented in SDL2, because the tearing doesn't occur when in a window,
even when that window is the same dimensions as the fullscreen
resolution. The Direct3D renderer doesn't exhibit this problem. By
default, SDL2 uses a renderer in the following order of preference:
Direct3D > OpenGL > Software. So now, I've cleared the vid_scaledriver
cvar, and Doom Retro will use Direct3D by default, falling to OpenGL if
DirectX isn't installed, and software if that fails. You do get less FPS
with Direct3D, but I much prefer a smoother experience.