Make: Fix for compilation regardless of platform
The prior commit "fixed" compilation for 64-bit Linux (actually it
didn't, because the Makefile used spaces instead of tabs...) by
changing the line from -m32 to -m64.  There is no need to do such
things, the toolchain on distributions knows well what the native
architecture is (and -m32/-m64 is gcc-specific anyway, making it kind
of moot for people that use other compilers like icc).  Also Doom
Retro's move to SDL2 wasn't considered in the original Makefile.

This should fix both issues and get rid of the linux64 file.