Implement improved screen scaling method
Thanks to some excellent work from @fabiangreffrath and @EMue. See the
discussion here:
https://github.com/chocolate-doom/chocolate-doom/issues/450. Default of
vid_scalefilter is now "best", with "linear" and "nearest" still
available. There are a few different versions of I_FinishUpdate() now,
which is a bit of a mess, but hopefully it keeps things as efficient as
possible.