Fix bleeding when using linear filter in Direct3D properly
Commit dd770f741249608cfb6bf49a80795f00d9362938 was a hack. Although it
seems it wasn't affecting other people, it was apparent on 2 notebooks I
tested on, and on an external monitor, at various resolutions. After a
bit of research, I've determined it's an issue with textures in Direct3D
not being clamped as they are in OpenGL. See a bug report here:
https://bugzilla.libsdl.org/show_bug.cgi?id=1878. Since SDL2 releases
are slow, I've compiled my own version of SDL2 with the patch described
in the above report applied, and the issue is now fixed.