Initial implementation of automap on a second display
There's a lot more work to do, but this commit will create a second
fullscreen window on a second display if available, and if the new
vid_automapdisplay cvar is on. When displayed, the tab key and all
automap controls are disabled. I still need to fix a focus problem at
startup, what to do between levels, display the map title correctly,
apply all vid_ cvars to the new window, etc., and probably lots more,
but at least its up and working.