Include SDL2_mixer build with madlib
It seems that SDL2_mixer's use of smpeg to play MP3s is shit, and is the
cause of issue #188. But after spending a considerable amount of time
compiling the libmad library (www.underbit.com/products/mad), and then
recompiling SDL2_mixer to use that instead of smpeg, MP3s appear to play
properly now.