Fix position of statbar nums when using statbar from PWAD
Since v1.0, the numbers in the status bar have been shifted down by 1px
so they are better centered vertically. However, in BTSX's status bar,
the % symbol is part of the STBAR lump rather than a separate lump, and
so makes the numbers out of alignment. This commit won't shift the
numbers down if a custom STBAR lump is found in the loaded PWAD.