Handle alternate config files differently
If an alternate config file is loaded (either by autoloading, selecting
it in the launcher, or by the -config cmd-line option), load
doomretro.cfg file first, then the alternate config, but still save to
doomretro.cfg and not overriding the settings changed in the alternate
config. Also, the configfile cvar has been removed.