OSX: save user files to Application Support instead of to the folder the app is in.
On OSX - and indeed on Linux and nowadays Windows - the convention is
to store generated application files (config files, logs, savegames
etc.) within a private user-specific folder, not within the location of
the application itself.

This commit implements this for OSX and adds suitable hooks for doing
so for Windows and Linux in future. It also standardises where
resources (currently just doomretro.wad) are loaded from, cleaning up
some OSX code forks.

I have not tested this on Windows, so it’s possible I’ve broken
something.