Display STARTUP5 string in Freedoom
There were 2 problems here. In DEHACKED lump, because there was no space
between "STARTUP5" and "=", the line wasn't parsed at all. Correcting
this, I had it so newlines were removed, which caused the string to be
printed all on line and truncated to fit.
Now STARTUP5 is parsed correctly, and is also split over multiple lines
in the console.