Correct defect in game restore of targets
From
https://github.com/jeffdoggett/Doom/commit/d727c7822d97f38d8ef3b645c8e85fb9ef07c9eb:
"The linked list of thinkers does not get restored in the same order
that it was when the game was saved. The P_MobjThinkers gets saved
first, followed by the other stuff. Similarly the P_MobjThinkers get
restored first, followed by the rest of the thinkers. So we can only
count the P_MobjThinkers when counting the position in the linked list."