Split gp_vibrate cvar into two
The `gp_vibrate` CVAR that toggles the vibration of *XInput* gamepads
has now been split into two CVARs: `gp_vibrate_damage` for when the
player is damaged, and `gp_vibrate_weapons` for when the player fires
their weapon. Both are `on` by default. Thanks to @L3gend for the
suggestion (issue #311).