Fix lighting of trees in boomedit.wad
A while ago I chose to use the floor light level to light all things in
a sector. This makes the trees in boomedit.wad too dark. So in the
instance of there being sky, the light level will be calculated as
before (average of ceiling and floor light levels).