Revamp use of episode/expansion/savegame/skilllevel cvars
Instead of being a number indicating the currently selected episode,
etc. in the menu, these cvars are now read-only strings of the episode,
etc. for the current game. (Currently selected episode, etc. still saved
as "episodeselected", etc. in doomretro.cfg, just not exposed in
console.) More work to do though.