Don't use pickup radius if thing moving vertically
The red skull key directly in front of the cyberdemon in MAP15 of Going
Down would get stuck in midair as the platform it is on is lowered. This
is due to code added to P_CheckPosition() in commit
f1f7a527667017f98aad4ed8a92a6cd5db01f56e. My fix is to not execute this
code if a thing is moving vertically. @jeffdoggett, this might be of
interest to you, (and you might be able to find a more elegant
solution!). Thanks to Tim Woodacre (Phobos Anomaly) for reporting this.