Implement vanilla mode!
"Vanilla mode" may now be toggled using the vanilla CCMD. This mode
changes 61 different CVARs, enabling and disabling features to make DOOM
Retro as close to Vanilla DOOM as possible. While this mode is enabled,
any changes to CVARs won't be saved, and so all CVARs will be restored
to their previous values once this mode is disabled.