Check all sectors player is touching are special
This fixes a vanilla bug that I haven't seen being mentioned before.
Consider a situation when the player is standing between two damaging
sectors of different heights (such as a few places in breach.wad).
Because the player is not considered _on_ the lower of the two sectors
until he drops down onto it, the player isn't damaged. This commit
checks each sector the player is touching and then calls
P_PlayerInSpecialSector() if any sector is both special and at the same
height of the player.