Treat IWAD as PWAD if official IWAD also selected
The wads for the Minor Sprite Fixing Project
(https://www.doomworld.com/idgames/graphics/sprfix18) are incorrectly
marked as IWADs in their headers. Other source ports generally don't
have an issue with this, but because DOOM Retro's WAD launcher relies on
this distinction to determine what's been selected and whether an IWAD
needs to be autoloaded first, these WADs exit with a "Flats section not
found in IWAD" error. This commit makes a change so that DR will first
search through the WADs selected for an official IWAD (DOOM.WAD or
DOOM2.WAD). If found, any other IWADs selected will be treated as PWADs.