Disable sprite offset fixes for BTSX
Now that the final release of BTSX incorporates the sprite fixes from
the Minor Sprite Fixing Project by Revenant100, they conflict with the
sprite offset fixes that DR applies itself. Thanks to Alia Immortalis
for the report. (Will need to also do this for when the Minor Sprite
Fixing Project separately...)