Disabled Code Sign On Copy
Xcode was not allowing me to build with the new SDL_Mixer and it was throwing an error about signing. Disabling "Code Sign On Copy" in the build phases resolved the error. I don't think this should have been enabled in the first place as these are 3rd party frameworks that are already signed.