Update player's view when on moving platform
Something that has always bothered me with Vanilla DOOM is this: when the player is on a lift, there's a distinct jolt of the player's view when said lift reaches its top or bottom. It seems the player's view height isn't updated as immediately as the floor height, causing the player to be slightly lower than they should when a lift is moving up, and slightly higher than they should when a lift is moving down. Test this on, for example, the lift in DOOM2.WAD's MAP01. Position yourself so you see part of the lift's floor and the zombieman that  you killed as the lift moves up and pay attention to how the player moves in relation to everything else. Hope I've explained this clearly. (BTW this happens in PrBoom+, but not in ZDoom.)