Attempt to fix accidental respawn with +fire action
DOOM Retro has a feature that allows you to repawn using the fire key/button upon death. There is a counter that is meant to stop you from accidentally respawn too soon, but Ryan Freeman has reported that in some cases it can trigger instantly. It turns out that this counter isn't reset in all instances, so hopefully this commit will fix this once and for all.