Fix [possible?] bug with P_InterceptVector()
This code was originally taken from PrBoom+, but looking at the original code in Vanilla Doom, and the code in ZDoom, it would appear that these values have been swapped. I can't be sure which is correct as I don't understand the math. More testing is required, and this could potentially be the cause of the melee problems in issue #446. Can anybody who actually reads these notes please shed some light on this?