Precache animated flats and textures
I noticed there were still some dropped frames occuring (see issue #476). It turns out this was occuring whenever an animated texture or flat was rendered on screen for the first time. The existing precaching code would handle the texture/flat the surface was initially set to, but not the remaining frames. This appears to fix those type of dropped frames.