Only allow armortype cvar to be changed when player has armor
I'm considering changing how armor works in the game. End result will be the same, but simplify how it is handled in the HUD and console. player_t::armortype is initialized to 1 when player spawned. It can never be 0. In P_DamageMobj(), instead of checking if amortype is non-zero, just check if armorpoints is non-zero. If armorpoints become zero after damage and armortype is blue, just make it green. If anyone is actually reading this, does it make sense?