Set iwadfolder to common DOOM installation on 1st start
Or at least, try to. Windows 7 and onwards will only allow initial folder of open file common dialog to be set progmatically when it is opened for the 1st time. I can't test this, because even separate installs of DR are counted as the same install dialog. Is someone with a clean install of DR, and (for example) a Steam install of D3BFG, able to confirm this works?