Implement autotilt feature
This feature was originally suggested by Jon Krazov quite a while ago, and I've finally got around to implementing it. It emulates a feature of Quake that causes the player's view to automatically tilt up/down when the player moves up/down stairs. It still requires a bit of work, but you get the idea. Maybe I'll figure out a better name for the CVAR as well...