Fix hang when stepping on line specials 154 and 240
This fixes issue #507. Thank you to @simc2 for the report! This was caused by calling P_ChangeSector() in EV_DoChange(), which I added a while ago to remove blood splats and clip/unclip sprites if a sector changes to/from liquid. Created a separate function to handle this, P_CheckLiquidSector().