Fix slight framerate drop when playing sfx
DR's performance has been as smooth as butter for a while now. But after a couple of minutes into starting a game, when an SFX is played, there is quite often a slight but noticeable drop in framerate. More testing is required, but I believe this may be the cause.