Update release notes
Crash can also occur whenever the player is firing a weapon and exits a map at the same time, so not just specials 197 and 198. I see it's because I failed to check if weapon state null in P_MovePsprites() when I refactored that function.  The fix I provided in the previous commit should be fine.