Change behavior of -nodeh slightly
Previously, -nodeh would stop loading of ALL dehacked lumps or files. Now, it will only stop the loading of dehacked lumps and the automatic loading of files. If a dehacked file is specifically selected in the launcher, or specified on the command-line using -deh, it will still be loaded. Thanks to Noiser for suggesting this.