Implement r_hudtype CVAR
The `r_althud` CVAR has been replaced with the new `r_hudtype` CVAR. Its value is the year of the version of *DOOM* the widescreen HUD is inspired by, which can be either `1993`, `2016` or `2020`. It is `1993` by default and when vanilla mode is on.

This is preparation for the 2020 HUD type inspired by DOOM Eternal...