Revamp random number generator
Revert to Vanilla DOOM's RNG for all things except the menu. (It's still seeded in the same way.) This was done because Vanilla DOOM's RNG doesn't have an exactly even distribution, so changing to an RNG that does affected damage, angle of gunshots, etc. for those with the eye for it. A big thanks goes again to Joshua Savarese for picking up on this subtle difference, and testing so we could get this right. Also see https://doomwiki.org/wiki/Pseudorandom_number_generator.