Clear unused flags when spawning thing
Lee Killough: "We clear the flags unused in DOOM if we see flag mask 256 set, since it is reserved to be 0 under the new scheme. A 1 in this reserved bit indicates it's a DOOM wad made by a DOOM editor which puts 1's in bits that weren't used in DOOM (such as HellMaker wads). So we should then simply ignore all upper bits."
Fixes #604. Thanks @P41R47!