Introduce r_nudge CVAR
The `r_corpses_nudge` CVAR has been deprecated and replaced with the `r_nudge` CVAR. When this CVAR is `on`, all corpses as well as all items that can be picked up by the player will now move slightly if walked over by the player or a monster. This CVAR is `on` by default and `off` when vanilla mode is enabled.