Show author name for one-humanity.wad in mapstats
https://romero.com/shop/p/onehumanity

"One Humanity is a new level for the 1994 id Software release DOOM® II created by John Romero to support the people of Ukraine and the humanitarian efforts of the Red Cross and the UN Central Emergency Response Fund. 100% of the proceeds go toward this support.

One Humanity is Romero’s first DOOM II level since the release of the original in 1994. The .WAD contains a readme text file as well as the external mod data. Players must have an original copy of DOOM II and a modern source port to play One Humanity."