Implement `english` CVAR
This is something I've wanted to do for a while. This CVAR toggles between using American and international English. For the moment, all this does is change "armor" to "armour" in the status bar, but will extend to all text output in the console and in player messages. I might even change "international" to "australian". "american" is default of course.