Implement `sucktime` CVAR
The amount of time in hours the player can spend in the current map before that time "SUCKS". Affects time in intermission screen, output of `am_mapstats` CVAR, and output of `playerstats` CCMD. `1` by default and in vanilla mode. Still working on this. May implement `sucktime` in `MAPINFO` lumps as well...