Attempt to fix SFX volume degradation issue
At least two users have reported this issue, and all I can put it down to is the fade out of all SFX applied when the menu or console is opened. Unable to replicate myself, but hopefully this change to S_FadeOutSounds() will fix it. If not, I'll just remove it entirely.