Don't try to install sprite lump if size is 0
This is what I've been spending a considerable amount of time tracking down. Running SD21.wad was giving me a R_InstallSpriteLump error. It's because of the empty markers used in the wad (the first one being call "TROO"). Dsdhacked support now appears to be working (and I almost gave up)!