Remove `direct3d11` support (again)
I've encountered another issue with SDL's Direct 3D 11 implementation, that does not occur when Direct 3D 9 is used. Best way to see this: go to the end of E1M1. Quickly approach and flick the exit switch and the screen will for some reason become a blend of two frames for a second before the wipe. (The other issue reported is corrupted screenshots, which I've been able replicate when taking a screenshot when using 2 displays and the `am_external` CVAR is `on`.)